#15 protocolをpresenterから分離
